+++ bin/service.dart
@@ -12,23 +12,23 @@ main(List<String> args) async {
 
   if (config?.servicesPath != null) {
     path = "${config.servicesPath}";
-    lib = "${config.servicesPath}/services.dart";
+    lib = "lib/services.dart";
   }
 
-  String dartPath = '$path/${toTableName(name)}_service.dart';
+  String prefix =  config?.servicesPath != null ? "lib/" : "";
+
+  String dartPath = '$prefix$path/${toTableName(name)}.dart';
 
   await writeInFile(dartPath, serviceTemplate(name));
 
   if (lib != null) {
-    addToLibrary("${toTableName(name)}_service.dart", lib);
+    addToLibrary("$path/${toTableName(name)}", lib);
   }
 
 }
 
 String serviceTemplate(String name) =>
-    '''// Copyright (c) 2016, <your name>. All rights reserved. Use of this source code
-// is governed by a BSD-style license that can be found in the LICENSE file.
-
+    '''
 import "package:angular2/core.dart";
 
 @Injectable()