+++ bin/template/lib/components/components.dart
@@ -1,3 +1,0 @@
-library __projectName__.components;
-
-export "home_component/home_component.dart";