+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.2.0
+* Add `init` command to create a config file inside an existing project
+* Update `Angular2`
+* move all libraries export to `lib/`
+* simplify generated file name (no '_component.dart', '_service.dart'...)
+* `sass` support
+* `less` support
+* refactor `Route` defintion
+
+
 ## 0.1.0
 
 * Add full application generation with `new` command